Fix review comments

+++ app/models/alias.rb
@@ -8,9 +8,9 @@ class Alias < ActiveRecord::Base
   validates :commit_name_id, presence: true
   validates :preferred_name_id, presence: true
 
-  after_save :schedule_project_analysis, if: :preferred_name_or_deleted_changed?
-  after_save :update_unclaimed_person, if: :created_or_deleted?
-  after_update :move_name_facts_to_preferred_name, if: :preferred_name_changed?
+  after_save :update_unclaimed_person, if: proc { id_changed? || deleted_changed? }
+  after_save :schedule_project_analysis, if: proc { preferred_name_id_changed? || deleted_changed? }
+  after_update :move_name_facts_to_preferred_name, if: proc { preferred_name_id_changed? }
 
   # TODO: Figure out explain_yourself
   acts_as_editable editable_attributes: [:preferred_name_id]
@@ -53,15 +53,15 @@ class Alias < ActiveRecord::Base
     end
 
     def create_for_project(editor_account, project, commit_name_id, preferred_name_id, override_permissions = false)
-      a = where(project_id: project.id, commit_name_id: commit_name_id).first_or_initialize
-      a.editor_account = editor_account
-      if a.persisted?
-        update_pre_existing_alias(editor_account, a, commit_name_id, preferred_name_id)
+      alias_obj = where(project_id: project.id, commit_name_id: commit_name_id).first_or_initialize
+      alias_obj.editor_account = editor_account
+      if alias_obj.persisted?
+        update_pre_existing_alias(editor_account, alias_obj, commit_name_id, preferred_name_id)
       else
-        a.assign_attributes(preferred_name_id: preferred_name_id)
-        override_permissions ? a.save_without_validation! : a.save!
+        alias_obj.assign_attributes(preferred_name_id: preferred_name_id)
+        override_permissions ? alias_obj.save_without_validation! : alias_obj.save!
       end
-      a
+      alias_obj
     end
 
     private
@@ -76,14 +76,14 @@ class Alias < ActiveRecord::Base
       aliases.join(projects).on(projects[:best_analysis_id].eq(aas[:analysis_id])).join_sources[0].to_sql
     end
 
-    def update_pre_existing_alias(editor_account, a, commit_name_id, preferred_name_id)
+    def update_pre_existing_alias(editor_account, alias_obj, commit_name_id, preferred_name_id)
       if commit_name_id == preferred_name_id
-        CreateEdit.where(target: a).first.undo!(editor_account) unless a.deleted
+        CreateEdit.where(target: alias_obj).first.undo!(editor_account) unless alias_obj.deleted
       else
-        CreateEdit.where(target: a).first.redo!(editor_account) if a.deleted
-        a.update_attributes(preferred_name_id: preferred_name_id)
+        CreateEdit.where(target: alias_obj).first.redo!(editor_account) if alias_obj.deleted
+        alias_obj.update_attributes(preferred_name_id: preferred_name_id)
       end
-      a.reload
+      alias_obj.reload
     end
   end
 
@@ -95,10 +95,11 @@ class Alias < ActiveRecord::Base
   def update_unclaimed_person
     person = Person.where(name_id: commit_name_id, project_id: project_id).first_or_create
     contributor_fact = ContributorFact.find_by(name_id: preferred_name_id, analysis_id: project.best_analysis_id)
+    return unless contributor_fact
     if deleted?
-      contributor_fact.try(:remove_name_fact, person.name_fact)
+      contributor_fact.remove_name_fact(person.name_fact)
     elsif person
-      contributor_fact.try(:append_name_fact, person.name_fact)
+      contributor_fact.append_name_fact(person.name_fact)
       person.destroy
     end
   end
@@ -109,17 +110,5 @@ class Alias < ActiveRecord::Base
       .try(:append_name_fact, name_fact)
     ContributorFact.find_by(name_id: preferred_name_id_was, analysis_id: project.best_analysis_id)
       .try(:remove_name_fact, name_fact)
-  end
-
-  def preferred_name_or_deleted_changed?
-    attribute_changed?(:preferred_name_id) || attribute_changed?(:deleted)
-  end
-
-  def preferred_name_changed?
-    attribute_changed?(:preferred_name_id)
-  end
-
-  def created_or_deleted?
-    attribute_changed?(:id) || attribute_changed?(:deleted)
   end
 end
